Create truncate-characters.js
Step 1 -
Select an element you want to truncate (may it be div, paragraph, text block). Give it a Class or a Combo class ‘target’.

Step 2 -
Go to > page settings > copy and paste the below code before the body tag as it is. You can modify the number of characters where it currently has ‘20’.